pyproject.toml
[tool.poetry]
name = "trading-bot-langgraph"
version = "1.0.0"
description = "변동성 돌파 자동매매 봇 (LangGraph 기반)"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    {include = "skills"},
    {include = "apps"}
]

[tool.poetry.dependencies]
python = "^3.9"
langgraph = "^0.2.0"
langchain-core = "^0.3.0"
pydantic = "^2.0.0"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"

# Flask 웹 앱
flask = "^3.0.0"

# 데이터 처리
pandas = "^2.0.0"
numpy = "^1.24.0"

# 로깅
colorlog = "^6.7.0"

# 백테스팅 (선택)
backtrader = "^1.9.78"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^24.0.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
trading-bot = "apps.daily_breakout_app:main"
trading-web = "apps.flask_app:main"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W"]
